reorganize unstructured k8s
