--- test/e2e/cni/unstructured.go
@@ -1,53 +1,0 @@
-package cni
-
-import (
-	"bufio"
-	"bytes"
-	"errors"
-	"fmt"
-	"io"
-
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	apiyaml "k8s.io/apimachinery/pkg/util/yaml"
-	"sigs.k8s.io/yaml"
-)
-
-// yamlToUnstructured takes a YAML and converts it to a list of Unstructured objects.
-func yamlToUnstructured(rawyaml []byte) ([]unstructured.Unstructured, error) {
-	var ret []unstructured.Unstructured
-
-	reader := apiyaml.NewYAMLReader(bufio.NewReader(bytes.NewReader(rawyaml)))
-	count := 1
-	for {
-		// Read one YAML document at a time, until io.EOF is returned
-		b, err := reader.Read()
-		if err != nil {
-			if errors.Is(err, io.EOF) {
-				break
-			}
-			return nil, fmt.Errorf("failed to read yaml: %w", err)
-		}
-		if len(b) == 0 {
-			break
-		}
-
-		var m map[string]interface{}
-		if err := yaml.Unmarshal(b, &m); err != nil {
-			return nil, fmt.Errorf("failed to unmarshal the %d yaml document: %q: %w", count, string(b), err)
-		}
-
-		var u unstructured.Unstructured
-		u.SetUnstructuredContent(m)
-
-		// Ignore empty objects.
-		// Empty objects are generated if there are weird things in manifest files like e.g. two --- in a row without a yaml doc in the middle
-		if u.Object == nil {
-			continue
-		}
-
-		ret = append(ret, u)
-		count++
-	}
-
-	return ret, nil
-}
